fix: ai generated page
